+++ api/kogito-services/src/test/java/org/kie/kogito/services/event/impl/DummyEvent.java
@@ -1,0 +1,19 @@
+package org.kie.kogito.services.event.impl;
+
+class DummyEvent {
+
+    private String dummyField;
+
+    @SuppressWarnings("unused")
+    public DummyEvent() {
+    }
+
+    public DummyEvent(String dummyField) {
+        this.dummyField = dummyField;
+    }
+
+    @SuppressWarnings("unused")
+    public String getDummyField() {
+        return dummyField;
+    }
+}
